requirements.txt
paho_mqtt==1.3.1
pyserial==3.4
PyYAML==3.12
